update project 2

+++ database/migrations/2024_09_23_134923_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 255);
             $table->string('image',255);
-            $table->decimal('price',10,3);
+            $table->integer('price');
             
             $table->integer('remaining_quantity');
             $table->string('description');
